Remove mount point for VB additions when no longer used
Previously this was set to always mount the VirtualBox additions on each
boot, this creates a problem if the additions ISO has been removed, it
is common with Vagrant boxes that the user home directory be free of any
other files than those required. As these are no longer required after
the initial install these should be removed. The removal is not within
the scope of this project but we shouldn't be mounting this on each boot
and breaking the machine